requirements.txt
pyserial == 2.6
pyFirmata == 1.0.3
